% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_days_obs.R
\name{calc_days_obs}
\alias{calc_days_obs}
\title{Calculate the number of days until the end of the
observation or test period}
\usage{
calc_days_obs(date_crossing, date_obs_end, dur_test)
}
\arguments{
\item{date_crossing}{A vector of dates specifying the dates of border
crossings. Can be of class \code{Date} or \code{character}, and must have
format yyyy-mm-dd. Cannot contain missing values.}

\item{date_obs_end}{The last date for which data are available.}

\item{dur_test}{The length of the test period.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}
}
\value{
An integer vector the same length as \code{date_crossing}.
}
\description{
Calculate, for each border crossing, the number of days
until the end of the observation period, or the
test period, whichever comes first.
}
\details{
\code{date_crossing} can contain dates after \code{date_obs_end}.
The return value for all such dates is 0.

The records should be in the same order that they are
for \code{\link{calc_days_in_country}}.
}
\examples{
date_crossing <- c("2009-10-12", "2010-01-05", "2009-03-01",
                   "2010-01-20", "2010-12-09")
date_obs_end <- "2010-06-30"
dur_test <- 480
calc_days_obs(date_crossing = date_crossing,
              date_obs_end = date_obs_end,
              dur_test = dur_test)
}
