% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_is_long_term_mig.R
\name{calc_is_long_term_mig}
\alias{calc_is_long_term_mig}
\title{Determine whether border crossing is long-term migration.}
\usage{
calc_is_long_term_mig(personId, is_arrival, res_status_before,
  res_status_after, days_in_country, days_obs, dur_test, dur_threshold,
  parallel = TRUE, n_core = 2, check = TRUE)
}
\arguments{
\item{personId}{A vector of person/passenger IDs.}

\item{is_arrival}{A logical vector, the same length as \code{personId}
and \code{date_crossing}, with each element specifying whether the
corresponding border crossing is an arrival.
For each person ID, the values of \code{is_arrival} must alternate between
\code{TRUE} and \code{FALSE}, starting with either
(depending on the direction of the first border crossing by the passenger).}

\item{res_status_before}{A vector of \code{0}s, \code{1}s, and \code{NA}s,
the same length as \code{personId}, giving residence status before
the border crossing.}

\item{res_status_after}{A vector of \code{0}s, \code{1}s, and \code{NA}s,
the same length as \code{personId}, giving residence status after
the border crossing.}

\item{days_in_country}{Number of days in the country
from the time of the border crossing to the end
of the observation or test periods. Usually calculated
with \code{\link{calc_days_in_country}}.}

\item{days_obs}{Number of days from the time of the border
crossing to the end of the observation period.
Usually calculated with \code{\link{calc_days_obs}}.}

\item{dur_test}{The length of the test period.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}

\item{dur_threshold}{The length of time a person must spend
in or out of the country, within the test period,
in order to change their residence status.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}

\item{parallel}{Logical. Whether to use parallel processing, to
speed up calculations.  Defaults to \code{TRUE}.}

\item{n_core}{The number of cores to use, if \code{parallel} is
\code{TRUE}. Defaults to \code{2}. Higher values will
typically result in faster calculations on computers
with more than two cores.}

\item{check}{Logical. Whether to perform sanity checks on
arguments. Defaults to \code{TRUE}.}
}
\value{
An integer vector the same length as \code{personId}.
}
\description{
Determine, for each border crossing, whether the crossing involves
a change in residence status, and therefore counts as a long-term
migration.  When there is insufficient information to decide
either way, the return value is set to missing.
}
\details{
The records should be in the same order that they are
for \code{\link{calc_days_in_country}}.
}
\examples{
personId <- c(1, 1, 1,
               2, 2,
               3, 3)
is_arrival <- c(FALSE, TRUE, FALSE,
                FALSE, FALSE,
                TRUE, FALSE)
res_status_before <- c(1, 0, 0,
                       1, NA,
                       0, 0)
res_status_after <- c(0, 0, 0,
                      NA, NA,
                      0, 0)
days_in_country <- c(20, 20, 0,
                     30, 30,
                     7, 0)
days_obs <- c(200, 150, 130,
              310, 30,
              70, 63)
dur_test <- 365
dur_threshold <- 300
calc_is_long_term_mig(personId = personId,
                      is_arrival = is_arrival,
                      res_status_before = res_status_before,
                      res_status_after = res_status_after,
                      days_in_country = days_in_country,
                      days_obs = days_obs,
                      dur_test = dur_test,
                      dur_threshold = dur_threshold)
}
