% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res_status_after.R
\name{calc_res_status_after}
\alias{calc_res_status_after}
\title{Calculate residence status after border crossing}
\usage{
calc_res_status_after(personId, is_arrival, days_in_country, days_obs,
  res_status_initial, dur_test, dur_threshold, parallel = TRUE,
  n_core = 2, check = TRUE)
}
\arguments{
\item{personId}{A vector of person/passenger IDs.}

\item{is_arrival}{A logical vector, the same length as \code{personId}
and \code{date_crossing}, with each element specifying whether the
corresponding border crossing is an arrival.
For each person ID, the values of \code{is_arrival} must alternate between
\code{TRUE} and \code{FALSE}, starting with either
(depending on the direction of the first border crossing by the passenger).}

\item{days_in_country}{Number of days in the country
from the time of the border crossing to the end
of the observation or test periods. Usually calculated
with \code{\link{calc_days_in_country}}.}

\item{days_obs}{Number of days from the time of the border
crossing to the end of the observation period.
Usually calculated with \code{\link{calc_days_obs}}.}

\item{res_status_initial}{A vector of \code{0}s and \code{1}s,
the same length as \code{personId}, giving each person's
residence status before their first border crossing.
If a person makes \code{n} border crossings, then that
person's residence status will be repeated \code{n} times
in \code{res_status_initial}.}

\item{dur_test}{The length of the test period.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}

\item{dur_threshold}{The length of time a person must spend
in or out of the country, within the test period,
in order to change their residence status.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}

\item{parallel}{Logical. Whether to use parallel processing, to
speed up calculations.  Defaults to \code{TRUE}.}

\item{n_core}{The number of cores to use, if \code{parallel} is
\code{TRUE}. Defaults to \code{2}. Higher values will
typically result in faster calculations on computers
with more than two cores.}

\item{check}{Logical. Whether to perform sanity checks on
arguments. Defaults to \code{TRUE}.}
}
\value{
An integer vector the same length as \code{personId}.
}
\description{
Calculate, for each border crossing, the passenger's
residence status after the crossing.  If insufficient
time has passed since the border crossing, it may not
yet be possible to calculate residence status. In this
case, the residence status is set to \code{NA}.
}
\details{
The records should be in the same order that they are
for \code{\link{calc_days_in_country}}.
}
\examples{
personId <- c(1, 1, 2, 2, 2, 3)
is_arrival <- c(TRUE, FALSE, FALSE, TRUE, FALSE, FALSE)
days_in_country <- c(10, 0, 390, 390, 365, 0)
days_obs <- c(200, 190, 500, 480, 475, 30)
res_status_initial <- c(0, 0, 1, 1, 1, 1)
dur_test <- 480
dur_threshold <- 365
calc_res_status_after(personId = personId,
                      is_arrival = is_arrival,
                      days_in_country = days_in_country,
                      days_obs = days_obs,
                      res_status_initial = res_status_initial,
                      dur_test = dur_test,
                      dur_threshold = dur_threshold)
}
