% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res_status_before.R
\name{calc_res_status_before}
\alias{calc_res_status_before}
\title{Calculate residence status before border crossing}
\usage{
calc_res_status_before(personId, res_status_initial, res_status_after,
  check = TRUE)
}
\arguments{
\item{personId}{A vector of person/passenger IDs.}

\item{res_status_initial}{A vector of \code{0}s and \code{1}s,
the same length as \code{personId}, giving each person's
residence status before their first border crossing.
If a person makes \code{n} border crossings, then that
person's residence status will be repeated \code{n} times
in \code{res_status_initial}.}

\item{res_status_after}{A vector of \code{0}s, \code{1}s, and \code{NA}s,
the same length as \code{personId}, giving residence status after
the border crossing.}

\item{check}{Logical. Whether to perform sanity checks on
arguments. Defaults to \code{TRUE}.}
}
\value{
An integer vector the same length as \code{personId}.
}
\description{
Calculate, for each border crossing, the passenger's
residence status before the crossing.  If insufficient
time has passed since the border crossing, it may not
yet be possible to calculate residence status. In this
case, the residence status is set to \code{NA}.
}
\details{
The records should be in the same order that they are
for \code{\link{calc_days_in_country}}.
}
\examples{
personId <- c(1, 1,
               2, 2, 2,
               3)
res_status_initial <- c(0, 0,
                             1, 1, 1,
                             1)
res_status_after <- c(1, 1,
                      1, 0, 0,
                      1)
calc_res_status_before(personId = personId,
                       res_status_initial = res_status_initial,
                       res_status_after = res_status_initial)
}
