% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mig_hist.R
\name{plot_mig_hist}
\alias{plot_mig_hist}
\title{Plot a migration history.}
\usage{
plot_mig_hist(date_crossing, is_arrival, res_status_before_str = NULL,
  res_status_after_str = NULL, date_first = NULL, date_last = NULL,
  show_dates = TRUE, show_days = TRUE, cex = 1, lwd = 1)
}
\arguments{
\item{date_crossing}{A vector of dates.}

\item{is_arrival}{A logical vector, the same length as \code{date_crossing}
specifying whether each border crossing is an arrival.}

\item{res_status_before_str}{Character or numeric vector,
the same length as \code{date_crossing}, showing residence
status before each crossing. Optional.}

\item{res_status_after_str}{Character or numeric vector,
the same length as \code{date_crossing}, showing residence
status after each crossing.}

\item{date_first}{The start date for the travel history. Optional.}

\item{date_last}{The end date for the travel history. Optional.}

\item{show_dates}{Logical. Whether to display the dates of each border
crossing.}

\item{show_days}{Logical. Whether to display the length, in days, of each
spell in or out of the country.}

\item{cex}{'Character expansion factor'. A number. Larger values
lead to larger text. Defaults to 1.}

\item{lwd}{Line width. A number. Larger values lead to thicker lines.
Defaults to 1.}
}
\value{
Returns \code{NULL}, but as a side effect draws a graph
(using R's traditional graphics system).
}
\description{
Given a sequence of border crossings for a person,
draw a diagram describing that person's migration
history.
}
\details{
Note that, unlike elsewhere in package \code{migstat}, the
\code{date_crossing} and \code{is_arrival} arguments for
\code{plot_mig_hist} refer to a single individual.

If values for \code{date_first} and \code{date_last} are not supplied,
then defaults are calculated, based on the length of the travel history.
}
\examples{
date_crossing <- as.Date(c("2000-01-01", "2000-02-05", "2000-03-01"))
is_arrival <- c(FALSE, TRUE, FALSE)
plot_mig_hist(date_crossing = date_crossing,
              is_arrival = is_arrival)
}
